fix: user can only delete their own posts

+++ src/use-case/delete-post.ts
@@ -1,5 +1,6 @@
 import { Post } from '@prisma/client';
 import { PrismaPostsRepository } from '@/repositories/prisma/posts-repository';
+import { PrismaUsersRepository } from '@/repositories/prisma/users-repository';
 
 interface DeletePostRequest {
     postID: string;
@@ -11,12 +12,24 @@ interface DeletePostResponse {
 }
 
 export class DeletePost {
-    constructor(private postsRepository: PrismaPostsRepository) {}
+    constructor(
+        private postsRepository: PrismaPostsRepository,
+        private userRepository: PrismaUsersRepository,
+    ) {}
 
     async deletePost(
         postRequest: DeletePostRequest,
     ): Promise<DeletePostResponse> {
         const { postID, userID } = postRequest;
+        const user = await this.userRepository.findById(userID);
+
+        if (!user) {
+            throw new Error('🔍 User not found');
+        }
+
+        if (user.id !== userID) {
+            throw new Error('🚧 Unauthorized access');
+        }
 
         const deletedPost = await this.postsRepository.delete(postID, userID);
